bootstrap: script tweaks and docs

+++ scripts/build.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
+START_TIME=$(date +%s)
+
 BUILD_DIR="bootstrap"
+mkdir -p "$BUILD_DIR"
 
-mkdir -p $BUILD_DIR # tmp dir for bootstrap
-
+COMPILER=""
 for cmd in clang gcc cc; do
     if command -v "$cmd" >/dev/null 2>&1; then
         COMPILER="$cmd"
@@ -12,11 +14,8 @@ for cmd in clang gcc cc; do
     fi
 done
 
-if ! $COMPILER -std=c23 -dM -E - < /dev/null > /dev/null 2>&1; then
-    echo "------------------------------------------------------------"
-    echo "ERROR: Prehistoric compiler: ($COMPILER) detected."
-    echo "Aborting!!"
-    echo "------------------------------------------------------------"
+if [ -z "$COMPILER" ]; then
+    echo "[build.sh]: Error: No C compiler found."
     exit 1
 fi
 
@@ -27,22 +26,59 @@ elif command -v ld.lld >/dev/null 2>&1; then
     L_STRAT="-fuse-ld=lld"
 fi
 
+ARGS=(
+    -std=c23
+    -O2
+    -fPIE
+    -I. 
+    -Isrc 
+    -Isrc/barr_build_system 
+    -Isrc/barr_commands 
+    -Isrc/barr_platform/linux 
+    -Isrc/olmos
+    -D_GNU_SOURCE 
+    -DNDEBUG 
+    -mavx
+    -o "$BUILD_DIR/barr"
+    -lxxhash 
+    -lpthread
+)
 
-echo [build.sh]: Using $COMPILER...
-$COMPILER -std=c23 -O2 -fPIE \
-    -I. -Isrc -Isrc/barr_build_system -Isrc/barr_commands -Isrc/barr_platform/linux -Isrc/olmos \
-    $(find src -name "*.c") \
-    -D_GNU_SOURCE -DNDEBUG -mavx \ 
-    -o bootstrap/barr -lxxhash -lpthread $L_STRAT
+SOURCES=$(find src -name "*.c")
 
-echo [build.sh]: Rebuilding with barr...
-echo 
+echo "[build.sh]: Using $COMPILER..."
 
-rm -rf .barr             # removing any previous barr stale state 
-./$BUILD_DIR/barr -I     # initializing the cmake created barr
-./$BUILD_DIR/barr -rb    # rebuild barr 
+$COMPILER "${ARGS[@]}" $SOURCES $L_STRAT &
+BARR_PID=$!
 
-echo [build.sh]: Done, cleaning up...
-rm -rf $BUILD_DIR 
+width=40
+printf "Building: ["
+while ps -p $BARR_PID > /dev/null; do
+    printf "━"
+    sleep 0.2
+done
+
+printf "\r\033[KBuilding: [━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━] 100%%\n"
+
+wait $BARR_PID
+
+if [ -f "./$BUILD_DIR/barr" ]; then
+    echo "[build.sh]: Rebuilding with barr..."
+    echo 
+    rm -rf .barr
+    ./$BUILD_DIR/barr -I
+    ./$BUILD_DIR/barr -rb
+    
+    echo "[build.sh]: Done, cleaning up..."
+    rm -rf "$BUILD_DIR"
+else
+    echo "[build.sh]: Error: Bootstrap failed."
+    exit 1
+fi
 
 
+END_TIME=$(date +%s)
+TOTAL_TIME=$((END_TIME - START_TIME))
+
+echo "--------------------------------------"
+echo "[build.sh]: Total process took ${TOTAL_TIME}s"
